requirements.txt
# opendiamond core
pip>=1.5.6
M2Crypto==0.25.1
Pillow==4.0.0
lxml==3.7.3
python-dateutil==2.6.0
six==1.10.0

# jsonblaster
file-magic==0.3.0
pycurl==7.43.0
simplejson==3.10.0
sockjs-tornado==1.0.3
tornado==4.4.2
validictory==1.1.0

# dataretriever
Paste==2.0.3
flickrapi==2.2.1

# diamondd
redis==2.10.5

# scopeserver
Django==1.5.12

# testing
flake8>=3.0.4
pylint>=1.5.6
pylint-django>=0.7.2
pytest>=2.9.2
pytest-timeout>=1.0.0
